[ai-assisted] refactor(react): move shared frontend boundaries
Why:
- Address issue #53 by removing React runtime imports from Vue-era or tsconfig-excluded shared locations.

What:
- Moved the AG Grid Korean locale into the React component boundary and left a legacy re-export for the old Vue-era path.
- Reimplemented document and object storage React APIs on the React API/query client boundary.
- Moved the forum role permission matrix data into the React forum admin boundary.
- Updated CHANGELOG.md for issue #53.

Validation:
- npm run typecheck: pass
- npm run lint: pass with existing warnings only
- npm run build: pass; Vite reported existing large chunk warning
