+++ src/react/pages/documents/api.ts
@@ -1,4 +1,5 @@
-import { documentApi } from "@/data/studio/mgmt/document";
+import { apiClient } from "@/react/api/client";
+import { apiRequest } from "@/react/query/fetcher";
 import type {
   DocumentBlock,
   DocumentBlockCreateRequest,
@@ -27,28 +28,74 @@ export interface DocumentTreeParams {
   includeDeleted?: boolean;
 }
 
+interface ApiEnvelope<T> {
+  data?: T;
+}
+
+const API_BASE = "/api/mgmt/documents";
+
+function unwrapPayload<T>(payload: ApiEnvelope<T> | T) {
+  if (
+    payload &&
+    typeof payload === "object" &&
+    !Array.isArray(payload) &&
+    "data" in payload
+  ) {
+    return payload.data as T;
+  }
+
+  return payload as T;
+}
+
+async function requestWithMeta<T>(url: string, params?: Record<string, unknown>) {
+  const response = await apiClient.get<ApiEnvelope<T> | T>(url, {
+    params,
+    withCredentials: true,
+  });
+
+  return {
+    data: unwrapPayload(response.data),
+    etag: response.headers?.etag as string | undefined,
+  };
+}
+
 export const reactDocumentsApi = {
   getLatest(documentId: number): Promise<DocumentBundleResponse> {
-    return documentApi.getLatest(documentId);
+    return requestWithMeta<DocumentVersionBundle>(`${API_BASE}/${documentId}`);
   },
   getVersion(documentId: number, versionId: number): Promise<DocumentBundleResponse> {
-    return documentApi.getVersion(documentId, versionId);
+    return requestWithMeta<DocumentVersionBundle>(
+      `${API_BASE}/${documentId}/versions/${versionId}`
+    );
   },
   listBlocksTree(documentId: number, params?: DocumentTreeParams): Promise<{ data: DocumentBlockNode[] }> {
-    return documentApi.listBlocksTree(documentId, params);
+    return apiRequest<DocumentBlockNode[]>(
+      "get",
+      `${API_BASE}/${documentId}/blocks/tree`,
+      { params }
+    ).then((data) => ({ data }));
   },
   listBlocksByVersion(
     documentId: number,
     versionId: number,
     params?: { includeDeleted?: boolean; parentBlockId?: number }
   ): Promise<DocumentBlocksResponse> {
-    return documentApi.listBlocksByVersion(documentId, versionId, params);
+    return requestWithMeta<DocumentBlock[]>(
+      `${API_BASE}/${documentId}/versions/${versionId}/blocks`,
+      params
+    );
   },
   getBlock(documentId: number, blockId: number): Promise<DocumentBlockResponse> {
-    return documentApi.getBlock(documentId, blockId);
+    return requestWithMeta<DocumentBlock>(
+      `${API_BASE}/${documentId}/blocks/${blockId}`
+    );
   },
   createBlock(documentId: number, payload: DocumentBlockCreateRequest) {
-    return documentApi.createBlock(documentId, payload);
+    return apiRequest<{ blockId: number }, DocumentBlockCreateRequest>(
+      "post",
+      `${API_BASE}/${documentId}/blocks`,
+      { data: payload }
+    );
   },
   updateBlock(
     documentId: number,
@@ -56,9 +103,18 @@ export const reactDocumentsApi = {
     payload: DocumentBlockUpdateRequest,
     ifMatch?: string
   ) {
-    return documentApi.updateBlock(documentId, blockId, payload, ifMatch);
+    return apiRequest<void, DocumentBlockUpdateRequest>(
+      "put",
+      `${API_BASE}/${documentId}/blocks/${blockId}`,
+      {
+        data: payload,
+        headers: ifMatch ? { "If-Match": ifMatch } : {},
+      }
+    );
   },
   deleteBlock(documentId: number, blockId: number, ifMatch?: string) {
-    return documentApi.deleteBlock(documentId, blockId, ifMatch);
+    return apiRequest<void>("delete", `${API_BASE}/${documentId}/blocks/${blockId}`, {
+      headers: ifMatch ? { "If-Match": ifMatch } : {},
+    });
   },
 };
